cloud/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.0
numpy==1.26.4
scikit-learn==1.4.1.post1
python-dotenv==1.0.1
requests==2.31.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
